fix: trivial change
playercomponents는 불필요한 클래스로 판단되어 제거
