+++ Gravity Controller/Assets/Prefabs/Player/Player.prefab
@@ -175,7 +175,6 @@ GameObject:
   - component: {fileID: 1577732291757072104}
   - component: {fileID: 6306873086425688456}
   - component: {fileID: 2008502258580397860}
-  - component: {fileID: 5096196705468182801}
   - component: {fileID: 8708367440453553676}
   - component: {fileID: 6477734520506417645}
   - component: {fileID: 9102416230853527869}
@@ -184,7 +183,6 @@ GameObject:
   - component: {fileID: 3647855855870178485}
   - component: {fileID: 8793847713277570272}
   - component: {fileID: 780792836414734017}
-  - component: {fileID: 1934565738}
   m_Layer: 0
   m_Name: Player
   m_TagString: Player
@@ -305,19 +303,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 3dbced07cee4a46e0bf9d33004774538, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!114 &5096196705468182801
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3940993412661136708}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 339aafc1101dc472aab60184416c7d16, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  _sparkParticle: {fileID: 4802001733177264264, guid: 91b5e47d130ae0641a81b111c88a0f4b, type: 3}
 --- !u!114 &8708367440453553676
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -366,6 +351,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d915936d3a82c4aaf9335c68cebcbc10, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  _sparkParticle: {fileID: 4802001733177264264, guid: 91b5e47d130ae0641a81b111c88a0f4b, type: 3}
   _isShootable: 1
   _isReloading: 0
 --- !u!114 &595114278171343580
@@ -405,102 +391,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!82 &780792836414734017
-AudioSource:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3940993412661136708}
-  m_Enabled: 1
-  serializedVersion: 4
-  OutputAudioMixerGroup: {fileID: 0}
-  m_audioClip: {fileID: 0}
-  m_PlayOnAwake: 1
-  m_Volume: 1
-  m_Pitch: 1
-  Loop: 0
-  Mute: 0
-  Spatialize: 0
-  SpatializePostEffects: 0
-  Priority: 128
-  DopplerLevel: 1
-  MinDistance: 1
-  MaxDistance: 500
-  Pan2D: 0
-  rolloffMode: 0
-  BypassEffects: 0
-  BypassListenerEffects: 0
-  BypassReverbZones: 0
-  rolloffCustomCurve:
-    serializedVersion: 2
-    m_Curve:
-    - serializedVersion: 3
-      time: 0
-      value: 1
-      inSlope: 0
-      outSlope: 0
-      tangentMode: 0
-      weightedMode: 0
-      inWeight: 0.33333334
-      outWeight: 0.33333334
-    - serializedVersion: 3
-      time: 1
-      value: 0
-      inSlope: 0
-      outSlope: 0
-      tangentMode: 0
-      weightedMode: 0
-      inWeight: 0.33333334
-      outWeight: 0.33333334
-    m_PreInfinity: 2
-    m_PostInfinity: 2
-    m_RotationOrder: 4
-  panLevelCustomCurve:
-    serializedVersion: 2
-    m_Curve:
-    - serializedVersion: 3
-      time: 0
-      value: 0
-      inSlope: 0
-      outSlope: 0
-      tangentMode: 0
-      weightedMode: 0
-      inWeight: 0.33333334
-      outWeight: 0.33333334
-    m_PreInfinity: 2
-    m_PostInfinity: 2
-    m_RotationOrder: 4
-  spreadCustomCurve:
-    serializedVersion: 2
-    m_Curve:
-    - serializedVersion: 3
-      time: 0
-      value: 0
-      inSlope: 0
-      outSlope: 0
-      tangentMode: 0
-      weightedMode: 0
-      inWeight: 0.33333334
-      outWeight: 0.33333334
-    m_PreInfinity: 2
-    m_PostInfinity: 2
-    m_RotationOrder: 4
-  reverbZoneMixCustomCurve:
-    serializedVersion: 2
-    m_Curve:
-    - serializedVersion: 3
-      time: 0
-      value: 1
-      inSlope: 0
-      outSlope: 0
-      tangentMode: 0
-      weightedMode: 0
-      inWeight: 0.33333334
-      outWeight: 0.33333334
-    m_PreInfinity: 2
-    m_PostInfinity: 2
-    m_RotationOrder: 4
---- !u!82 &1934565738
 AudioSource:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
