--- Gravity Controller/Assets/Scripts/Player/PlayerComponents.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 339aafc1101dc472aab60184416c7d16
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
